pyproject.toml
[tool.poetry]
name = "nlp-workshop"
version = "0.1.0"
description = "NLP Tech Training"
authors = ["Ingo Marquart <ingo.marquart@statworx.com>"]

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.18.0"
tokenizers = "^0.12.1"
torch = "^1.11.0"
torchvision = "^0.12.0"
pandas = "^1.4.2"
numpy = "^1.22.3"
huggingface-hub = "^0.5.1"
bitsandbytes-cuda111 = "^0.26.0"
datasets = "^2.1.0"
fastpunct = "^2.0.2"
ipywidgets = "^7.7.0"
editdistance = "^0.6.0"
nltk = "^3.7"
spacy = "^3.3.0"
dask = "^2022.5.0"
keybert = "^0.5.1"
phonemizer = "^3.1.1"
pytorch-lightning = "^1.6.3"
torchtext = "^0.12.0"
torchdata = "^0.3.0"
category-encoders = "^2.4.1"
s3fs = "^2022.3.0"
scikit-learn = "^1.1.0"
matplotlib = "^3.5.2"
bertviz = "^1.4.0"
bert-serving-client = "^1.10.0"
gensim = "^4.2.0"
jupyter-contrib-nbextensions = "^0.5.1"

[tool.poetry.dev-dependencies]
darglint = "^1.8.1"
flake8 = "^4.0.1"
flake8-docstrings = "^1.6.0"
isort = "^5.10.1"
mypy = "^0.950"
pep8-naming = "^0.12.1"
pydocstyle = "^6.1.1"
black = "^22.3.0"
ipykernel = "^6.13.0"
ipython = "^8.3.0"
pre-commit = "^2.19.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
